fix bugs; heristic method for org group

+++ craftsman/transformer_manager.py
@@ -80,6 +80,7 @@ class TransformerManager(object):
         defs.ORDER_WHEN = order_when
         defs.EXPRIMENT_COL = expriment_col
         defs.EXPRIMENT_METHOD = expriment_col_stragey
+        defs.GROUP = group
 
         model = load_model(model_file)
         pipeline_features_in = model.feature_names_in_.tolist()
@@ -94,34 +95,42 @@ class TransformerManager(object):
         min_cost = float("inf")
 
         if group == 'org':
-            # enumerate the chain implement plans
-            all_chain_candidate_implement_plans: list[ChainCandidateImplementPlans] = []
-            for feature, chain in preprocessing_graph.chains.items():
-                all_chain_candidate_implement_plans.append(ChainCandidateImplementPlans(feature, chain))
+            reuse_flag = False
+            if reuse_flag and os.path.exists(f'{table_name}_{model_name}_org.pkl'):
+                with open(f'{table_name}_{model_name}_org.pkl', 'rb') as f:
+                    min_cost_preprocessing_graph = pickle.load(f)
+                    
+            else:
+                # enumerate the chain implement plans
+                all_chain_candidate_implement_plans: list[ChainCandidateImplementPlans] = []
+                for feature, chain in preprocessing_graph.chains.items():
+                    all_chain_candidate_implement_plans.append(ChainCandidateImplementPlans(feature, chain))
 
-            def enumerate_graph_implement_plans(index=0, current_graph_implement_plan:list = []):
-                if index == len(all_chain_candidate_implement_plans):
-                    yield copy.deepcopy(current_graph_implement_plan) 
-                    return
-                for chain_implement_plan in all_chain_candidate_implement_plans[index].candidate_implement_plans:
-                    current_graph_implement_plan.append(chain_implement_plan)
-                    yield from enumerate_graph_implement_plans(index + 1, current_graph_implement_plan)
-                    current_graph_implement_plan.pop()
-            
-            for graph_implement_plan in enumerate_graph_implement_plans():
-                preprocessing_graph = implement_operator_by_plan(preprocessing_graph, graph_implement_plan)
-                if cost_model == 'craftsman':
-                    cost = get_craftsman_graph_cost(preprocessing_graph, data_rows)
-                elif cost_model == 'postgresql':
-                    query_str = self.__compose_sql(join_the_operators(preprocessing_graph), table_name, dbms, pre_sql, pipeline)
-                    cost = get_pg_sql_cost(query_str)
-                if cost < min_cost:
-                    min_cost_preprocessing_graph = preprocessing_graph
-                    min_cost = cost
+                def enumerate_graph_implement_plans(index=0, current_graph_implement_plan:list = []):
+                    if index == len(all_chain_candidate_implement_plans):
+                        yield copy.deepcopy(current_graph_implement_plan) 
+                        return
+                    for chain_implement_plan in all_chain_candidate_implement_plans[index].candidate_implement_plans:
+                        current_graph_implement_plan.append(chain_implement_plan)
+                        yield from enumerate_graph_implement_plans(index + 1, current_graph_implement_plan)
+                        current_graph_implement_plan.pop()
+                
+                for graph_implement_plan in enumerate_graph_implement_plans():
+                    preprocessing_graph = implement_operator_by_plan(preprocessing_graph, graph_implement_plan)
+                    if cost_model == 'craftsman':
+                        cost = get_craftsman_graph_cost(preprocessing_graph, data_rows)
+                    elif cost_model == 'postgresql':
+                        query_str = self.__compose_sql(join_the_operators(preprocessing_graph), table_name, dbms, pre_sql, pipeline)
+                        cost = get_pg_sql_cost(query_str)
+                    if cost < min_cost:
+                        min_cost_preprocessing_graph = preprocessing_graph
+                        min_cost = cost
+                        
+                with open(f'{table_name}_{model_name}_org.pkl', 'wb') as f:
+                    pickle.dump(min_cost_preprocessing_graph, f)
                     
             min_cost_query_str = self.__compose_sql(join_the_operators(min_cost_preprocessing_graph), table_name, dbms, pre_sql, pipeline)      
-            with open(f'{table_name}_{model_name}_org.pkl', 'wb') as f:
-                pickle.dump(min_cost_preprocessing_graph, f)
+            
 
         if group == 'pos':
             if os.path.exists(f'{table_name}_{model_name}_org.pkl'):
@@ -279,8 +288,6 @@ class TransformerManager(object):
                 all_chain_candidate_fusion_plans,
                 preprocessing_graph.chains.keys(),
             ):
-                if feature == 'Brand':
-                    pass
                 for chain_implement_plan in chain_candidate_implement_plans.candidate_implement_plans:
                     for chain_fusion_plan in chain_candidate_fusion_plans.candidate_fusion_plans:
                         preprocessing_graph_list = merge_sql_operator_by_chain_plan(last_chain_min_cost_preprocessing_graph, chain_implement_plan, chain_fusion_plan, feature)
@@ -292,9 +299,6 @@ class TransformerManager(object):
                                 query_str = self.__compose_sql(graph, table_name, dbms, pre_sql, pipeline)
                                 cost = get_pg_sql_cost(query_str)
                             if cost < min_cost:
-                                if cost_model == 'craftsman':
-                                    query_str = self.__compose_sql(graph, table_name, dbms, pre_sql, pipeline)
-                                min_cost_query_str = query_str
                                 min_cost_preprocessing_graph = graph
                                 min_cost = cost
                 
@@ -349,7 +353,7 @@ class TransformerManager(object):
             join_feature_list[op.features[0]] = join_features
 
         # compose imputer sql if exists missing cols
-        if pipeline['imputer']['missing_cols']:
+        if pipeline['imputer']['missing_cols'] or join_feature_sqls:
             filled_values = pipeline['imputer']['filled_values']
             missing_cols = pipeline['imputer']['missing_cols']
             missing_col_indexs = pipeline['imputer']['missing_col_indexs']
